requirements.txt
python-dotenv
docling
pandas
streamlit
opencv-python
pillow
google-generativeai
dashscope
openai
ipykernel
einops
ultralytics
timm
paddlepaddle
paddleocr
molscribe
torch
torchvision
timm
numpy<2.0
